pyproject.toml
[project]
name = "searchgov"
version = "0.1.0"
description = "Multimodal search for singapore data"
authors = [{ name = "Syamil Maulod", email = "syamil8@gmail.com" }]
requires-python = ">=3.11.5,<3.12"
readme = "README.md"
license = "MIT"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "rapidfuzz>=3.9.4,<4",
    "pandas>=2.2.2,<3",
    "faiss-cpu>=1.9.0,<2",
    "polars>=1.11.0,<2",
    "loguru>=0.7.2,<0.8",
    "beautifulsoup4>=4.12.3,<5",
    "sqlite-vec>=0.1.6,<0.2",
    "sf-hamilton[visualization]>=1.85.1,<2",
    "paramiko>=3.5.0,<4",
    "fabric>=3.2.2,<4",
    "duckdb>=1.2.2,<2",
    "psycopg2-binary>=2.9.10,<3",
    "einops>=0.8.1,<0.9",
    "numpy<2",
    "pgvector>=0.4.1,<0.5",
    "thefuzz>=0.22.1,<0.23",
    "nicegui>=2.19.0,<3",
    "python-dotenv>=1.1.0,<2",
    "pyecharts>=2.0.8,<3",
    "python-lsp-server>=1.12.2,<2",
    "websockets>=15.0.1,<16",
    "python-lsp-ruff>=2.2.2,<3",
    "supabase>=2.13.0",
    "pydantic>=1.10.22,<2",
    "asyncpg>=0.30.0",
    "fastparquet>=2024.11.0",
    "pyarrow>=21.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.5.1,<0.6",
    "pytest>=8.2.2,<9",
    "ipykernel>=6.29.5,<7",
    "matplotlib>=3.9.1,<4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
